requirements.txt
openai==0.28.0
streamlit
streamlit_chat
langchain
sentence_transformers
pinecone-client